[Gold V] Title: 자연수 색칠하기, Time: 20 ms, Memory: 8884 KB -BaekjoonHub

+++ 백준/Gold/23048. 자연수 색칠하기/자연수 색칠하기.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int colors[500'001];
+unsigned short colors[500'001];
 
 int main() {
   ios_base::sync_with_stdio(false);
